Extract transform logic from `ColliderHierarchyPlugin` to `ColliderTransformPlugin`
